Reorganize themes, add careers-site initiative
This reorganizes the themes for the team to recognize that we maintain a
few projects now that are not Taskcluster (MLS, Pulseguardian, Careers),
and operate a bit of those too.  Some of the themes remain TC-specific,
but some (data and costs) are more general.  Note that I moved work on
making Taskcluster more operations-friendly into the `usability` theme.

This also adds a new "careers-website" initiative to track work on that
website.